Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
sqlalchemy==2.0.23
sqlmodel==0.0.25
python-dotenv==1.0.1
cryptography==46.0.1
python-multipart==0.0.3
PyJWT == 2.10.0
passlib == 1.7.4